requirements.txt
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pdf2image>=1.16.0
python-docx>=1.1.0
